fix: Sucessfully changed to Typescript, and got page up and running again.
