+++ .github/workflows/builddeploy.yaml
@@ -3,7 +3,7 @@ name: Deploy React to GitHub Pages
 on:
   push:
     branches:
-      - typescript # name of the branch you are pushing to
+      - main # name of the branch you are pushing to
 
 jobs:
   build:
